requirements.txt
python-dotenv-1.0.1
gradio-4.43.0
langchain-0.2.16
langchain-community-0.2.16
openai-1.37.0
